Minor fixes in comm size events
git-svn-id: http://compss.bsc.es/svn/compss/framework/trunk@2254 9ab3861e-6c05-4e1b-b5ef-99af60850597

+++ compss/runtime/adaptors/nio/commons/src/main/java/integratedtoolkit/nio/NIOTracer.java
@@ -71,10 +71,10 @@ public class NIOTracer extends Tracer {
         emitCommEvent(send, partnerID, tag, 0);
     }
 
-    public static void emitCommEvent(boolean send, int partnerID, int tag, int size){
+    public static void emitCommEvent(boolean send, int partnerID, int tag, long size){
 
         synchronized(Tracer.class){
-            Wrapper.Comm(send, tag, size, partnerID, ID);
+            Wrapper.Comm(send, tag, (int)size, partnerID, ID);
         }
 
         if (debug) {
